Medicine-Stock-Microservice/src/main/java/com/cognizant/medicinestock/model/ErrorResponse.java
package com.cognizant.medicinestock.model;

import java.time.LocalDateTime;

import org.springframework.http.HttpStatus;

/**
 * This is a model class for custom error handling which which have fields like
 * time stamp, status, reason and a message.
 */
public class ErrorResponse {

	private LocalDateTime timestamp;
	private HttpStatus status;
	private String reason;
	private String message;

	public ErrorResponse() {
	}

	public LocalDateTime getTimestamp() {
		return timestamp;
	}

	public void setTimestamp(LocalDateTime timestamp) {
		this.timestamp = timestamp;
	}

	public HttpStatus getStatus() {
		return status;
	}

	public void setStatus(HttpStatus status) {
		this.status = status;
	}

	public String getReason() {
		return reason;
	}

	public void setReason(String reason) {
		this.reason = reason;
	}

	public String getMessage() {
		return message;
	}

	public void setMessage(String message) {
		this.message = message;
	}
}